Introduction
Bike sharing system is the system in which its features like membership, renting a bike and return back is automatic which is the convenient way to allow users to rent a bike easily from some specific place and return it at different place. Now-a-days these systems are in a great use as they are playing important role in traffic as an efficient transport system, contributes to environmental and some health problems due to its involvement to healthy, active and lifestyle. We know that about over 500 bike-sharing systems are existing worldwide that is made up over 500 thousand bicycles. The attributes of data created by these systems is interesting which make them best for the research and motivate researchers to observe and work on this data. There are other transport services like buses or subways, light rails; bike sharing systems can combine these modes of transport, stores the information related to these services and give users manifold choices in their journey. Hence these features turn and work as a virtual sensor network which is used to sense the movement of the automobiles in the public. So, in this way it is possible to identify the events by observing this data.